+++ joinfaces-autoconfigure/src/main/java/org/joinfaces/autoconfigure/omnifaces/OmnifacesInitializerRegistrationBean.java
@@ -14,40 +14,27 @@
  * limitations under the License.
  */
 
-package org.joinfaces.autoconfigure.mojarra;
+package org.joinfaces.autoconfigure.omnifaces;
 
 import java.util.Set;
 
 import javax.servlet.annotation.HandlesTypes;
 
-import com.sun.faces.config.FacesInitializer;
-import org.joinfaces.autoconfigure.JsfClassFactory;
-import org.joinfaces.autoconfigure.JsfClassFactoryConfiguration;
 import org.joinfaces.autoconfigure.ServletContainerInitializerRegistrationBean;
+import org.omnifaces.facesviews.FacesViewsInitializer;
 
 /**
- * Servlet Context Initializer of Mojarra.
- *
+ * Servlet context initializer of OmniFaces.
  * @author Marcelo Fernandes
  */
-public class MojarraServletContextInitializer extends ServletContainerInitializerRegistrationBean<FacesInitializer> {
+public class OmnifacesInitializerRegistrationBean extends ServletContainerInitializerRegistrationBean<FacesViewsInitializer> {
 
-	/**
-	 * Constant of another faces config of mojarra.
-	 */
-	public static final String ANOTHER_FACES_CONFIG = "com/sun/faces/jsf-ri-runtime.xml";
-
-	public MojarraServletContextInitializer() {
-		super(FacesInitializer.class);
+	public OmnifacesInitializerRegistrationBean() {
+		super(FacesViewsInitializer.class);
 	}
 
 	@Override
 	protected Set<Class<?>> getClasses(HandlesTypes handlesTypes) {
-		JsfClassFactory jsfClassFactory = new JsfClassFactory(JsfClassFactoryConfiguration.builder()
-				.anotherFacesConfig(ANOTHER_FACES_CONFIG)
-				.handlesTypes(handlesTypes)
-				.excludeScopedAnnotations(true)
-				.build());
-		return jsfClassFactory.getAllClasses();
+		throw new IllegalStateException(FacesViewsInitializer.class + " is not annotated with " + HandlesTypes.class);
 	}
 }
